requirements.txt

Django==5.2.12
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-yasg==1.21.15
opencv-python==4.13.0.92
pillow==12.1.1
python-decouple==3.8
tensorflow_cpu==2.21.0
whitenoise==6.12.0

